Start of refactoring the configuration lib
Aligning the library with other libraries such as
authenticate and session so they all work in
similar ways.
