+++ library/configuration/ftrack/library/configuration/helper/metadata.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class METADATA(Enum):
+    ROOT = "_metadata"
+    SOURCES = "sources"
+    DELETE = "delete-after-compose"
+    CONFLICTS = "conflicts"
